+++ ChangeLog.rst
@@ -2,7 +2,15 @@
 ChangeLog
 =========
 
-0.1.0
-=====
+0.2.0 (2016-08-14)
+==================
+
+* Added graceful error handling.
+* Plus, github token now can be specified either
+  via ``--token`` option or through ``TOKEN`` environment
+  variable.
+
+0.1.0 (2016-08-13)
+==================
 
 Initial version, released a command line tool.